finish finger extraction

+++ object_detection/utils.py
@@ -56,7 +56,7 @@ def locate_hand_landmarks(image_path: str, detector_path: str) -> vision.HandLan
     # Convert the input image to a mediapipe image
     mediapipe_image = mp.Image.create_from_file(image_path)
     # Use the HandLandmarker object to detect hand landmarks in the mediapipe image
-    return mediapipe_image.numpy_view(), detector.detect(mediapipe_image)
+    return mediapipe_image.numpy_view().astype(np.uint8), detector.detect(mediapipe_image)
 
 
 def draw_landmarks_on_image(rgb_image: np.ndarray, detection_result: vision.HandLandmarkerResult) -> np.ndarray:
